creates first color from main.ts
