+++ src/components/ColorTitle.svelte
@@ -6,7 +6,7 @@
 
 <div class="flex flex-row items-center justify-between w-full max-w-screen-sm">
   <h3 class="text-gray-400 text-center">
-    Current color is {getColorString(color)}
+    {color ? `Current color is ${getColorString(color)}` : 'Loading...'}
   </h3>
   <div>
     <svg
@@ -15,7 +15,11 @@
       height={50}
       xmlns="http://www.w3.org/2000/svg"
     >
-      <rect height={50} width={50} fill={getColorString(color)} />
+      <rect
+        height={50}
+        width={50}
+        fill={color ? getColorString(color) : 'none'}
+      />
     </svg>
   </div>
 </div>
