+++ src/main.ts
@@ -1,9 +1,10 @@
 import App from './App.svelte'
+import { getRandomColor } from './utils'
 
 const app = new App({
   target: document.body,
   props: {
-    name: 'world'
+    color: getRandomColor()
   }
 })
