--- src/utils/constants.ts
@@ -1,1 +1,0 @@
-export const SVGN = 'http://www.w3.org/2000/svg'
