+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './num'
-export * from './constants'
 export * from './getDimensions'
 export * from './getColors'
 export * from './getColorString'
